+++ ingeniamotion/communication.py
@@ -642,7 +642,27 @@ class Communication(metaclass=MCMetaClass):
         )
 
     @staticmethod
+    def scan_servos_ethercat_with_info(
+        interface_name: str,
+    ) -> OrderedDict[int, SlaveInfo]:
+        """Scan a network adapter to get all connected EtherCAT slaves including slave information.
+
+        Args:
+            interface_name : interface name. It should have format
+                ``\\Device\\NPF_[...]``.
+
+        Returns:
+            Dictionary of nodes available in the network and slave information.
+
+        Raises:
+            TypeError: If some parameter has a wrong type.
+        """
+        net = EthercatNetwork(interface_name)
+        slaves_info = net.scan_slaves_info()
+        return slaves_info
+
     def scan_servos_ethercat(
+        self,
         interface_name: str,
     ) -> List[int]:
         """Scan a network adapter to get all connected EtherCAT slaves.
@@ -658,8 +678,6 @@ class Communication(metaclass=MCMetaClass):
         """
         net = EthercatNetwork(interface_name)
         slaves = net.scan_slaves()
-        if not isinstance(slaves, List):
-            raise TypeError("Slaves are not saved in a List")
         return slaves
 
     def scan_servos_ethercat_interface_ip(self, interface_ip: str) -> List[int]:
@@ -689,6 +707,42 @@ class Communication(metaclass=MCMetaClass):
         """
         return self.scan_servos_ethercat(self.get_ifname_by_index(if_index))
 
+    def scan_servos_canopen_with_info(
+        self,
+        can_device: CAN_DEVICE,
+        baudrate: CAN_BAUDRATE = CAN_BAUDRATE.Baudrate_1M,
+        channel: int = 0,
+    ) -> OrderedDict[int, SlaveInfo]:
+        """Scan CANOpen device network to get all nodes including slave information.
+
+        Args:
+            can_device : CANOpen device type.
+            baudrate : communication baudrate. 1 Mbit/s by default.
+            channel : CANOpen device channel. ``0`` by default.
+
+        Returns:
+            Dictionary of nodes available in the network and slave information.
+
+        Raise:
+            TypeError: If some parameter has a wrong type.
+        """
+        net_key = f"{can_device}_{channel}_{baudrate}"
+        if net_key not in self.mc.net:
+            self.mc.net[net_key] = CanopenNetwork(can_device, channel, baudrate)
+        net = self.mc.net[net_key]
+
+        if net is None:
+            self.logger.warning(
+                "Could not find any nodes in the network."
+                "Device: %s, channel: %s and baudrate: %s.",
+                can_device,
+                channel,
+                baudrate,
+            )
+            return []
+        slaves_info = net.scan_slaves_info()
+        return slaves_info
+
     def scan_servos_canopen(
         self,
         can_device: CAN_DEVICE,
@@ -701,8 +755,11 @@ class Communication(metaclass=MCMetaClass):
             can_device : CANOpen device type.
             baudrate : communication baudrate. 1 Mbit/s by default.
             channel : CANOpen device channel. ``0`` by default.
+
         Returns:
             List of node ids available in the network.
+
+        Raises:
             TypeError: If some parameter has a wrong type.
         """
         net_key = f"{can_device}_{channel}_{baudrate}"
@@ -720,8 +777,6 @@ class Communication(metaclass=MCMetaClass):
             )
             return []
         slaves = net.scan_slaves()
-        if not isinstance(slaves, List):
-            raise TypeError("Slaves are not saved in a List")
         return slaves
 
     def disconnect(self, servo: str = DEFAULT_SERVO) -> None: